% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.lmbreak.R
\name{as.lmbreak}
\alias{as.lmbreak}
\title{From Multiple Breakpoint models to a Single Breakpoint Model}
\usage{
as.lmbreak(object, cluster)
}
\arguments{
\item{object}{output of \code{\link{mlmbreak}}.}

\item{cluster}{[numeric or character] cluster relative to which the breakpoint model should be converted.}
}
\description{
Retrive the output of \code{\link{lmbreak}} from \code{\link{mlmbreak}} for a specific cluster.
}
\keyword{utilities}
