% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.lmbreak}
\alias{autoplot.lmbreak}
\title{Graphical Display of the Breakpoint Model}
\usage{
\method{autoplot}{lmbreak}(
  object,
  xlim = NULL,
  ylim = NULL,
  color = grDevices::palette.colors(2),
  size = c(1.5, 2.5),
  linewidth = 1,
  alpha = 1,
  title = NULL,
  ...
)
}
\arguments{
\item{object}{output of \code{lmbreak}}

\item{xlim}{[numeric vector of length 2] range of displayed values for the x-axis (breakpoint variable).}

\item{ylim}{[numeric vector of length 2] range of displayed values for the y-axis (response variable).}

\item{color}{[character vector of length 2] colors used to display the observations and the model fit.}

\item{size}{[numeric vector of length 2] size of the points representing the observations and size of the points representing the breakpoints.}

\item{linewidth}{[numeric] width of the line representing the model fit}

\item{alpha}{[numeric] transparency of the points representing the observations. Can be set to 0 to hide the observations and only display the fitted broken lines.}

\item{title}{[character] character string to be displayed at the top of the graphical window.}

\item{...}{not used. For compatibility with the generic function.}
}
\value{
A list containing a ggplot2 object (element \code{plot}) and the dataset with the fitted values (element \code{data})
}
\description{
Display the fitted response value as a function of the breakpoint values
}
