% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.lmbreak}
\alias{coef.lmbreak}
\title{Extract Summary Statistics From Breakpoint Model}
\usage{
\method{coef}{lmbreak}(object, type = "breakpoint", simplify = TRUE, ...)
}
\arguments{
\item{object}{output of \code{\link{lmbreak}}.}

\item{type}{[vector of character] summary statistic to be output: \itemize{
\item breakpoint: position of the breakpoint
\item breakpoint.range: position of the breakpoint, minimum and maximum X values.
\item duration: duration (in term of X values) of each broken line.
\item Us: slope terms in the linear regression model used to estimates the parameters.
\item slope: slope associated to each breakpoint.
\item Vs: discontinuity terms in the linear regression model used to estimates the parameters.
\item intercept: expected outcome value at each breakpoint value.
\item lm: regression coefficients of the linear regression model used to estimates the parameters.
}}

\item{simplify}{[logical] simplify the data format from a list to a vector or a data.frame.}

\item{...}{Not used. For compatibility with the generic method.}
}
\value{
A vector when a single type is requested. A list or a data.frame otherwise.
}
\description{
Extract summary statistics from the breakpoint model
}
\keyword{methods}
