% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.mlmbreak}
\alias{coef.mlmbreak}
\title{Extract Summary Statistics From Multiple Breakpoint Model}
\usage{
\method{coef}{mlmbreak}(object, type = "breakpoint", cluster = NULL, format = "data.frame", ...)
}
\arguments{
\item{object}{output of \code{\link{mlmbreak}}.}

\item{type}{[character vector] summary statistic to be output. See argument \code{type} in \code{\link{coef.lmbreak}}.}

\item{cluster}{[vector] cluster relative to which the summary statistics should be extracted..}

\item{format}{[character] should the output be a data.frame (with cluster as a column) or a list (with one element for each cluster)?}

\item{...}{additional arguments passed to \code{\link{coef.lmbreak}}.}
}
\value{
A data.frame or a list.
}
\description{
Extract summary statistics from each breakpoint model
}
\keyword{methods}
