% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{SDIpsilo}
\alias{SDIpsilo}
\title{SDI data after Psilocibin}
\usage{
data(SDIpsilo)
}
\description{
Data from an experimental study where 16 healthy subjects receve an oral dose of psilocibin and their SDI (subjective drug intensity) score is followed over time.

\itemize{
\item \code{id}: study participant.
\item \code{time}: time (in minutes) since psilocibin intake.
\item \code{datapoint}: \code{"signal"} or \code{"noise"}. Indicates when the SDI measurement was done just after the participant visited the toilet which substantively, but temporarily, disrupted the experience (noise)
\item \code{score}: SDI score between 0 and 10
}
}
\references{
Stenbæk DS, Madsen MK, Ozenne B, et al. Brain serotonin 2A receptor binding predicts subjective temporal and mystical effects of psilocybin in healthy humans. Journal of Psychopharmacology. 2021;35(4):459-468. doi:10.1177/0269881120959609.
}
\keyword{datasets}
