% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmbreak.options.R
\name{lmbreak.options}
\alias{lmbreak.options}
\title{Global options for lmbreak package}
\usage{
lmbreak.options(..., reinitialise = FALSE)
}
\arguments{
\item{...}{options to be selected or updated}

\item{reinitialise}{should all the global parameters be set to their default value}
}
\value{
A list containing the default options.
}
\description{
Update or select global options for the lmbreak package.
}
\details{
The options are: \itemize{
\item n.iter [integer, >0] the maximum number of iterations used to estimates the breakpoints. Used by \code{lmbreak}.
\item tol [numeric, >0] the maximum accpetable difference between two consecutive estimates of the breakpoints. Used by \code{lmbreak}.
\item enforce.continuity [logical]: enforce.continuity [logical] in the case where no continuous solution could be found,
should a non-continuous breakpoint model be kept (\code{FALSE}) or a continuous breakpoint model be enforced (\code{TRUE}, refiting without the Vs terms). Used by \code{lmbreak}.
}
}
\keyword{utilities}
