% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedbreak.R
\name{logLik_mixed10}
\alias{logLik_mixed10}
\title{Mixed Model Likelihood for 10 pattern}
\usage{
logLik_mixed10(
  Y,
  t,
  beta,
  psi,
  tau_uu,
  tau_uv,
  tau_vv,
  sigma2,
  lower = -5,
  upper = 5,
  indiv = TRUE,
  integration
)
}
\arguments{
\item{Y}{[numeric vector] outcome values}

\item{t}{[numeric vector] timepoints at which the outcome value were measured.}

\item{beta}{[numeric] value of the population slope parameter.}

\item{psi}{[numeric] value of the population breakpoint.}

\item{tau_uu}{[numeric,>0] variance of the subject specific deviations to the population slope.}

\item{tau_uv}{[numeric,>0] covariance between the subject specific deviations slope and breapoint.}

\item{tau_vv}{[numeric] variance of the subject specific deviations to the population breakpoint.}

\item{sigma2}{[numeric,>0] residual variance.}

\item{lower}{[numeric] lower bound for the numerical integration: below this value the density of the random effect distribution is ignored.}

\item{upper}{[numeric] upper bound for the numerical integration: above this value the density of the random effect distribution is ignored.}

\item{indiv}{[logical] should the log-likelihood be output for each observation or only its sum across all observations}

\item{integration}{[character] method use to compute the likelihood: \code{"bivariate"} or \code{"univariate"}.
The latter is much faster.}
}
\value{
A numeric value for the value of the total log-likelihood (\code{indiv=FALSE}) or a \code{data.frame} with as many rows as observations (\code{indiv=TRUE}).
}
\description{
Evaluate the log-likelihood of a breakpoint model with 1 random slope, 1 random breakpoint, and one plateau (10 pattern) using numerical integration.
}
\examples{
#### EXAMPLE 1 ####
logLik_mixed10(Y = 0, t = 0, beta = 0, psi = 0, tau_uu = 1, tau_uv = 0, tau_vv = 1, sigma2 = 1, lower = -5, upper = 5, integration = "univariate")
##   integral1       error1 integral2       error2  integral error    logLik
##    1.253313 1.086413e-09 0.9896668 6.379719e-06 0.3569814    NA -1.030072

## first integral
sqrt(2*pi) * pnorm(5)*(1-pnorm(0))
## second integral
pnorm(0) * integrate(f = function(u){return(exp(-u^2/2) / sqrt(1+u^2))}, lower = -5, upper = 5)$value

logLik_mixed10(Y = 0, t = 0, beta = 0, psi = 0, tau_uu = 1, tau_uv = 0, tau_vv = 1, sigma2 = 1, lower = -5, upper = 5, integration = "bivariate")
## integral1 error1 integral2 error2  integral        error    logLik
##        NA     NA        NA     NA 0.3569812 3.522105e-06 -1.030072

#### EXAMPLE 2 ####
logLik_mixed10(Y = 0, t = 0, beta = 1, psi = 1, tau_uu = 1, tau_uv = 0, tau_vv = 1, sigma2 = 1, lower = -5, upper = 5, integration = "univariate")
##    integral1       error1 integral2       error2 integral error     logLik
## 1  2.108937 1.828095e-09  0.227924 7.136171e-07 0.371923    NA -0.9890684

logLik_mixed10(Y = 0, t = 0, beta = 1, psi = 1, tau_uu = 1, tau_uv = 0, tau_vv = 1, sigma2 = 1, lower = -5, upper = 5, integration = "bivariate")
##  integral1 error1 integral2 error2  integral        error     logLik
##         NA     NA        NA     NA 0.3862079 3.845569e-06 -0.9513795
}
