% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmbreak.R
\name{mlmbreak}
\alias{mlmbreak}
\title{Fit Multiple Breakpoint models}
\usage{
mlmbreak(formula, data, cluster, trace = 1, cpus = 1, ...)
}
\arguments{
\item{formula}{a formula where the breakpoint variable appears on the right hand side
with an argument pattern specifying the number of breakpoints and possible constrains. See details section of \code{\link{lmbreak}}.}

\item{data}{[data.frame] dataset}

\item{cluster}{[character] variable in the dataset indicating the cluster level.}

\item{trace}{[0,1,2,3] trace the execution of the function.}

\item{cpus}{[integer, >0] the number of CPU to use. When greater than 1, calculations are done parallel for different clusters.}

\item{...}{additional arguments passed to \code{\link{lmbreak}}.}
}
\description{
Fit a linear regression with breakpoints separately on each cluster of data.
Observations are independent between cluster but can be correlated within cluster.
}
\examples{

####  simulate data ####
set.seed(10)
df1 <- simBreak(c(10,25), breakpoint = c(0,1.5,3,4,5.5), slope = c(1,0,-1.2,0), sigma = 0.05)

#### fit breakpoint regression ####
e.mlmbreak1010 <- mlmbreak(Y ~ 0 + bp(X, "1010"), cluster = "id", data = df1)
plot(e.mlmbreak1010, scales = "free")
plot(e.mlmbreak1010, scales = "free", subtitle = 0)

## inspect results for a specific breakpoint
e4.lmbreak1010 <- as.lmbreak(e.mlmbreak1010, cluster = 4)
summary(e4.lmbreak1010)

## re-run with adpative step 
e4.lmbreak1010 <- lmbreak(Y ~ 0 + bp(X, "1010"), data = df1[df1$id==4,],
                          control = list(optimize.step = TRUE))
plot(e4.lmbreak1010)
}
\seealso{
\code{\link{as.lmbreak}} to extract the results specific to a single cluster of data. \cr
\code{\link{lmbreak}} for fitting the breakpoint model on a single cluster of data. \cr
\code{\link{model.tables.mlmbreak}} for extracting the breakpoints, slopes, intercept, and duration. \cr
\code{\link{plot.mlmbreak}} for a graphical display of the fitted breakpoint model. \cr
}
\keyword{models}
