% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.tables.R
\name{model.tables.lmbreak}
\alias{model.tables.lmbreak}
\title{Extract Key Summary Statistics From Breakpoint Model}
\usage{
\method{model.tables}{lmbreak}(x, ...)
}
\arguments{
\item{x}{output of \code{\link{lmbreak}}}

\item{...}{Not used. For compatibility with the generic method.}
}
\description{
Extract key summary statistics from the breakpoint model
}
\keyword{methods}
