% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.tables.R
\name{model.tables.mlmbreak}
\alias{model.tables.mlmbreak}
\title{Extract Key Summary Statistics From Multiple Breakpoint Model}
\usage{
\method{model.tables}{mlmbreak}(x, cluster = NULL, format = "data.frame", ...)
}
\arguments{
\item{x}{output of \code{\link{mlmbreak}}.}

\item{cluster}{[numeric or character] cluster relative to which the summary statistics should be extracted.}

\item{format}{[character] should the output be a data.frame (with cluster as a column) or an array (with cluster as the third dimension)?}

\item{...}{Passed to \code{model.tables.lmbreak}.}
}
\description{
Extract key summary statistics from each breakpoint model
}
\keyword{methods}
