% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R, R/plot.R
\name{plot.mlmbreak}
\alias{plot.mlmbreak}
\alias{autoplot.mlmbreak}
\title{Graphical Display for Multiple Breakpoint Model}
\usage{
\method{autoplot}{mlmbreak}(
  object,
  cluster = NULL,
  y = NULL,
  xlim = NULL,
  ylim = NULL,
  color = grDevices::palette.colors(3),
  size = c(1.5, 2.5),
  linewidth = 1,
  alpha = 1,
  scales = "fixed",
  labeller = "label_both",
  subtitle = 2,
  ...
)

\method{plot}{mlmbreak}(x, y = NULL, ...)
}
\arguments{
\item{object, x}{[mlmbreak] output of \code{\link{mlmbreak}}.}

\item{cluster}{[vector] cluster relative to which the breakpoint model should be displayed.}

\item{y}{[mlmbreak] another model whose fit is to be compared with the first one.}

\item{xlim}{[numeric vector of length 2] range of displayed values for the x-axis (breakpoint variable).}

\item{ylim}{[numeric vector of length 2] range of displayed values for the y-axis (response variable).}

\item{color}{[character vector of length 2 or 3] colors used to display the observations and the model fit.}

\item{size}{[numeric vector of length 2] size of the points representing the observations and size of the points representing the breakpoints.}

\item{linewidth}{[numeric] width of the line representing the model fit}

\item{alpha}{[numeric] transparency of the points representing the observations. Can be set to 0 to hide the observations and only display the fitted broken lines.}

\item{scales}{[character] argument passed to \code{ggplot2::facet_wrap} when displaying covariate specific model fit.}

\item{labeller}{[character] function used to label each facet. Passed to \code{ggplot2::facet_wrap}.}

\item{subtitle}{[logical] should the subtitle include information about convergence of the model.}

\item{...}{not used. For compatibility with the generic function.}
}
\value{
A list containing a ggplot2 object (element \code{plot}) and the dataset with the fitted values (element \code{data})
}
\description{
Display the fitted response value for each breapoint model as a function of the breakpoint values
}
\keyword{hplot}
