% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedbreak.R
\name{logLik_mixed10}
\alias{logLik_mixed10}
\title{Mixed Model Likelihood for 10 pattern}
\usage{
logLik_mixed10(
  Y,
  t,
  beta,
  psi,
  tau_uu,
  tau_uv,
  tau_vv,
  sigma2,
  lower = -5,
  upper = 5,
  indiv = TRUE,
  integration = "univariate-fast"
)
}
\arguments{
\item{Y}{[numeric vector] outcome values}

\item{t}{[numeric vector] timepoints at which the outcome value were measured.}

\item{beta}{[numeric] value of the population slope parameter.}

\item{psi}{[numeric] value of the population breakpoint.}

\item{tau_uu}{[numeric,>0] variance of the subject specific deviations to the population slope.}

\item{tau_uv}{[numeric,>0] covariance between the subject specific deviations slope and breapoint.}

\item{tau_vv}{[numeric] variance of the subject specific deviations to the population breakpoint.}

\item{sigma2}{[numeric,>0] residual variance.}

\item{lower}{[numeric] lower bound for the numerical integration: below this value the density of the random effect distribution is ignored.}

\item{upper}{[numeric] upper bound for the numerical integration: above this value the density of the random effect distribution is ignored.}

\item{indiv}{[logical] should the log-likelihood be output for each observation or only its sum across all observations}

\item{integration}{[character] method use to compute the likelihood: \code{"bivariate"} or \code{"univariate"}.
The latter is much faster. \code{"univariate-fast"} precomputes some quantities involved in \code{"univariate"} so is slightly faster.}
}
\value{
A numeric value for the value of the total log-likelihood (\code{indiv=FALSE}) or a \code{data.frame} with as many rows as observations (\code{indiv=TRUE}).
}
\description{
Evaluate the log-likelihood of a breakpoint model with 1 random slope, 1 random breakpoint, and one plateau (10 pattern) using numerical integration.
}
\examples{
#### EXAMPLE 1 ####
logLik_mixed10(Y = 0, t = 0, beta = 0, psi = 0, tau_uu = 1, tau_uv = 0, tau_vv = 1, sigma2 = 1, lower = -5, upper = 5, integration = "univariate")
##   integral1      error1 integral2       error2  integral error    logLik
## 1  0.199471 1.72908e-10 0.1575104 1.015364e-06 0.3569814    NA -1.030072

## first integral
pnorm(5)*(1-pnorm(0)) / sqrt(2*pi)
## second integral
pnorm(0) * integrate(f = function(u){return(exp(-u^2/2) / sqrt(1+u^2))}, lower = -5, upper = 5)$value / (2*pi)

logLik_mixed10(Y = 0, t = 0, beta = 0, psi = 0, tau_uu = 1, tau_uv = 0, tau_vv = 1, sigma2 = 1, lower = -5, upper = 5, integration = "bivariate")
## integral1 error1 integral2 error2  integral        error    logLik
##        NA     NA        NA     NA 0.3569812 3.522105e-06 -1.030072

#### EXAMPLE 2 ####
logLik_mixed10(Y = 0, t = 0, beta = 1, psi = 1, tau_uu = 1, tau_uv = 0, tau_vv = 1, sigma2 = 1, lower = -5, upper = 5, integration = "univariate")
##   integral1       error1  integral2       error2  integral error     logLik
## 1 0.3356478 2.909504e-10 0.05056137 4.592356e-06 0.3862092    NA -0.9513762

logLik_mixed10(Y = 0, t = 0, beta = 1, psi = 1, tau_uu = 1, tau_uv = 0, tau_vv = 1, sigma2 = 1, lower = -5, upper = 5, integration = "bivariate")
##  integral1 error1 integral2 error2  integral        error     logLik
##         NA     NA        NA     NA 0.3862079 3.845569e-06 -0.9513795

#### EXAMPLE 3 ####
logLik_mixed10(Y = 0.5, t = 2.2, beta = 1, psi = 3, tau_uu = 2, tau_uv = 0.3, tau_vv = 0.25, sigma2 = 1, lower = -5, upper = 5, integration = "univariate")
##    integral1       error1   integral2       error2  integral error    logLik
## 1 0.09976617 1.909237e-06 0.007462216 6.232982e-09 0.1072284    NA -2.232794

logLik_mixed10(Y = 0.5, t = 2.2, beta = 1, psi = 3, tau_uu = 2, tau_uv = 0.3, tau_vv = 0.25, sigma2 = 1, lower = -5, upper = 5, integration = "bivariate")
##   integral1 error1 integral2 error2  integral        error    logLik
## 1        NA     NA        NA     NA 0.1072284 1.061953e-06 -2.232794
}
