% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mlmbreak}
\alias{predict.mlmbreak}
\title{Fitted Values of the Multiple Breakpoint Model}
\usage{
\method{predict}{mlmbreak}(
  object,
  newdata = NULL,
  cluster = NULL,
  extrapolate = FALSE,
  keep.newdata = TRUE,
  ...
)
}
\arguments{
\item{object}{[mlmbreak] output of \code{\link{mlmbreak}}.}

\item{newdata}{[data.frame] dataset containing the covariate to condition on when evaluated the expected outcome.}

\item{cluster}{[vector] cluster relative to which the fit of the breakpoint model should be output.}

\item{extrapolate}{[logical] should fitted values before the first or beyond the last observation be set to NA?}

\item{keep.newdata}{[logical] should the dataset be added as additional columns in the output?}

\item{...}{additional arguments passed to \code{predict.lmbreak}.}
}
\description{
Extract fitted values from each breakpoint model.
}
