% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mlmbreak}
\alias{predict.mlmbreak}
\title{Fitted Values of the Multiple Breakpoint Model}
\usage{
\method{predict}{mlmbreak}(object, cluster = NULL, ...)
}
\arguments{
\item{object}{[mlmbreak] output of \code{\link{mlmbreak}}.}

\item{cluster}{[numeric or character] cluster relative to which the fit of the breakpoint model should be output.}

\item{...}{additional arguments (newdata, extrapolate, continuity, keep.newdata) passed to \code{predict.lmbreak}.}
}
\description{
Extract fitted values from each breakpoint model.
}
